Vector Usage

+++ Twitch/Engine/Engine.h
@@ -13,7 +13,7 @@ class Engine {
 public:
 	static int SCREEN_WIDTH;
 	static int SCREEN_HEIGHT;
-	static double getDt();
+	static float getDt();
 
 	Engine();
 	~Engine();
@@ -28,8 +28,8 @@ public:
 
 private:
 	static GLFWwindow* window;
-	static double dt; //differce between now and last time.
-	double lastTime;
+	static float dt; //differce between now and last time.
+	float lastTime;
 };
 
 #endif
